requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Security-Too==4.1.2
Flask-WTF==0.15.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Pillow==8.3.1
email-validator==1.1.3